# RFC: XZ compress HBRT image