@williamspatrick 2016-09-15
@bofferdn, @e-liner  - The secureboot code added an extra layer to the PNOR VFS.  If we add additional sections to be XZ compressed, I wonder if we should try to integrate that feature directly into that new secureboot VFS layer.  With this proposed patch from Stewart we end up with effectively duplicated code.

On the performance aspect, I thought we were going to attempt to overlap PAYLOAD and HBRT loading with OCC start.  Did that ever happen?