@dcrowell77 2016-09-15
No performance tweaks have been done to load pnor partitions in the background.