@dcrowell77 2016-09-15
RE: Pushing decompression code into the VFS/PNOR code - I'm in total agreement with that idea as a better design.  The question is always about resources and schedule.  Putting this in as-is produces an immediate benefit (a few seconds) for very little work, while finding time to put the right solution in could take awhile.