@e-liner 2016-09-15
This looks okay, but I'm going to want to test a few images on it. Plus we'll need an additional op-build change to compress the image if we're going this route.  I'm running short on time this week, but I can try to look at it tomorrow.