@stewart-ibm 2016-09-15
Patrick Williams <notifications@github.com> writes:
> @bofferdn, @e-liner  - The secureboot code added an extra layer to the
> PNOR VFS.  If we add additional sections to be XZ compressed, I wonder
> if we should try to integrate that feature directly into that new
> secureboot VFS layer.  With this proposed patch from Stewart we end up
> with effectively duplicated code.

Hence the RFC, because that's totally what should happen rather than
have inconsistent XZ calls scattered throughout hostboot.

This patch should totally not go in as-is, even though the error
handling (which I totally removed) is eerily similar to the error
handling already loading the HBRT image.

Ultimately, it looks like we don't even have to load the HBRT image
until opal_prd starts (there's no way I can think of how to do this on
P8 without breaking ABI though, but we could go that way for P9).

All of the partitions read in one go could be XZ compressed pretty
easily and save a bunch of boot time.

> On the performance aspect, I thought we were going to attempt to
> overlap PAYLOAD and HBRT loading with OCC start.  Did that ever
> happen?

Tackling OCC start was the next thing to look at. ISTEP 21.1 was really
constrained by PNOR access and OCC start time when I delved into it,
with absolutely nothing else going on there except that, so everything
there (except loading the PAYLOAD itself) is a candidate for the PAYLOAD
to do in parallel with other tasks.

-- 
Stewart Smith
OPAL Architect, IBM.