@stewart-ibm 2016-09-15
Dan <notifications@github.com> writes:
> RE: Pushing decompression code into the VFS/PNOR code - I'm in total
> agreement with that idea as a better design.  The question is always
> about resources and schedule.  Putting this in as-is produces an
> immediate benefit (a few seconds) for very little work, while finding
> time to put the right solution in could take awhile.

I'm happy to do that work, may need some assistance with a few pointers
here and there.

After previous experiments, it seems like the best time to load the
PAYLOAD is during bringing up DRAM, as that's easiest the longest time
spent doing anything (especially with P8 and DDR4).. I know there's memory
constraints on larger systems (Brazos), but we probably have a bit more
leeway with the OpenPOWER kind of size... even if we only loaded a
subset of the PAYLOAD that early.

I know that if I implemented that it'd make all the IPL flow documents
out of date, but technical documentation that's up to date and current
just seems.... not quite right :)

-- 
Stewart Smith
OPAL Architect, IBM.