@stewart-ibm 2016-09-15
e-liner <notifications@github.com> writes:
> This looks okay, but I'm going to want to test a few images on
> it. Plus we'll need an additional op-build change to compress the
> image if we're going this route.  I'm running short on time this week,
> but I can try to look at it tomorrow.

I think there's a few more partitions we could compress.

WINK and OCC come to mind, and they compress fairly well.

-- 
Stewart Smith
OPAL Architect, IBM.