@stewart-ibm 2016-09-16
Dan <notifications@github.com> writes:
>> Question: does hostboot need to load the hbrt image? Is there anything stopping us loading it in skiboot? Is there any initial setup done?.
>
> We load it and point at it in the devtree because that was what we
> were told to do originally by the OPAL team.  Remember that we do this
> for both OpenPOWER and the enterprise systems.  In the latter case,
> OPAL needs to request the data from the FSP as a lid since you don't
> have access to the PNOR.  It would not be difficult to have divergent
> behavior on our side if OPAL can handle it though.

It was certainly the easy way to get things going, no problem to revisit
though :)

> However, one more wrinkle in this story is Secureboot.  Hostboot has
> to load HBRT during IPL because we are the only ones would can verify
> the signature.  I think that forces us to keep the current design
> where HBRT is pushed into mainstore as part of the Hostboot IPL.

We have to verify signatures in skiboot, so this shouldn't be too much
of an issue.

-- 
Stewart Smith
OPAL Architect, IBM.