@williamspatrick 2016-09-16
The "other payload" requires Hostboot to load HBRT.  Their secureboot support will be part of the HBRT image so they end up with a (chicken-egg) otherwise.