@stewart-ibm 2016-09-16
Patrick Williams <notifications@github.com> writes:
> The "other payload" requires Hostboot to load HBRT.  Their secureboot
> support will be part of the HBRT image so they end up with a
> (chicken-egg) otherwise.

Okay, I thought that was a possibility.

Seeing as there's existing logic based on the payload type, I guess that
could be used.

Architecturally though, it does seem neater to have less knowledge of
the payload implementation and fewer if (payload=blah)....

-- 
Stewart Smith
OPAL Architect, IBM.