@dcrowell77 2016-09-16
> I know there's memory constraints on larger systems (Brazos), but we probably have a bit more leeway with the OpenPOWER kind of size

Actually, we are significantly more memory constrained in the OP systems because we have a lot more code loaded.  We are very close to the edge during the memory initialization steps, that is typically where we have seen OOM issues crop up.  If you make any changes, be sure to test on a Garrison box with full memory, that has been our worst case scenario in the past.